Implement reproducible inputs stage (#14)

+++ src/buffalo_weight/feature_calculators/perimeter.py
@@ -3,5 +3,5 @@ from __future__ import annotations
 from buffalo_weight.feature_calculators.context import FeatureContext
 
 
-def calculate_perimeter(ctx: FeatureContext) -> int:
+def calculate_perimeter(ctx: FeatureContext) -> float:
     return ctx.perimeter
